requirements.txt
flask
flask_admin
flask_sqlalchemy
mysql-connector-python
exchangelib
python-dotenv
gitpython